d3d12 work
Lots of work on the renderer
Now in a working state with bare minimum resource types
placeholder implementation of root description and pipline state object

+++ Engine/renderer_d3d12/memory/first_fit_allocator.h
@@ -14,7 +14,7 @@ inline size_t range_size(void *begin, void *end) {
 }
 
 inline void *ptr_align(void *ptr, unsigned align) {
-	return reinterpret_cast<void*>((reinterpret_cast<size_t>(ptr) + (align - 1)) & ~(align - 1));
+	return reinterpret_cast<void*>((reinterpret_cast<size_t>(ptr) + (size_t)((align - 1)) & ~(align - 1)));
 }
 
 inline void *ptr_subtract(void *ptr1, void *ptr2) {
